Moved method to common folder
