+++ response_operations_ui/common/remove_whitespace_from_survey.py
@@ -1,0 +1,2 @@
+def remove_whitespace_from_survey_name(survey):
+    return survey.replace(' ', '')
